MIRTjim.jl Documentation

Contents

MIRTjim.jl Documentation

Contents

Overview

Index

Functions

Overview

Most results from the Michigan Image Reconstruction Toolbox (MIRT) are images that one will want to display.

This module exports a function jim (jiffy image display) for convenient display of a grayscale image.

Index

MIRTjim.MIRTjim

MIRTjim.jim

MIRTjim.jim

MIRTjim.jim

MIRTjim.jim

MIRTjim.jim

MIRTjim.jim

MIRTjim.jim

MIRTjim.nothing_else

Functions

MIRTjim.MIRTjim — Module

`MIRTjim` exports the "jiffy image display" `jim`

for the Michigan Image Reconstruction Toolbox (MIRT)

MIRTjim.jim — Method

jim(z, title::String ; kwargs...)

MIRTjim.jim — Method

jim(z, ...)

A jiffy image display of z using heatmap

in

z image, can be 2D or higher, if higher then it uses mosaicviews

option

aspect_ratio; default :equal

clim; default (minimum(z),maximum(z))

color (colormap, e.g. :hsv); default :grays

colorbar (e.g. :none); default :legend

ncol for mosaicview for 3D and higher arrays; default 0 does auto select

padval padding value for mosaic view; default minimum(z)

line3plot lines around sub image for 3d mosaic; default false

line3type line type around sub image for 3d mosaic; default (:yellow)

mosaic_npad # of pixel padding for mosaic view; default 1

fft0 if true use FFTView to display; default false

title; default ""

xlabel; default ""

ylabel; default ""

yflip; default true if minimum(y) >= 0

yreverse; default true if y[1] > y[end]

x values for x axis; default collect(axes(z)[1])

y values for y axis; default collect(axes(z)[2])

xticks; default [minimum(x),maximum(x)]

yticks; default [minimum(y),maximum(y)]

out

returns plot handle, type Plots.Plot

2019-02-23 Jeff Fessler, University of Michigan

MIRTjim.jim — Method

jim(x, y, z, title::String ; kwargs...)

MIRTjim.jim — Method

jim(x, y, z ; kwargs...)

MIRTjim.jim — Method

jim(key::Symbol, value::Any)

set default value for one of the keys

MIRTjim.jim — Method

jim(:test::Symbol)

jim(:keys) return default keys

jim(:defs) return Dict of default keys / vals

jim(:key) return Dict[key] if possible

jim(:blank) return blank plot

MIRTjim.jim — Method

jim()

return docstring if user calls jim() with no arguments

MIRTjim.nothing_else — Method

nothing_else(x, y)

return y if x is nothing, else return x